requirements.txt
uvicorn==0.32.0
fastapi==0.115.3
asyncpg==0.30.0
python-dotenv==1.0.1
loguru==0.7.2
